requirements.txt
argcomplete==3.5.3
asgiref==3.8.1
build==1.2.2.post1
CacheControl==0.14.2
certifi==2025.1.31
charset-normalizer==3.4.1
cleo==2.1.0
click==8.1.8
colorama==0.4.6
crashtest==0.4.1
distlib==0.3.9
Django==5.2.1
django-cors-headers==4.7.0
django-rest-framework==0.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dulwich==0.22.7
fastjsonschema==2.21.1
filelock==3.17.0
gunicorn==23.0.0
idna==3.10
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
keyring==25.6.0
more-itertools==10.6.0
msgpack==1.1.0
packaging==24.2
pipx==1.7.1
pkginfo==1.12.0
platformdirs==4.3.6
poetry==2.0.1
poetry-core==2.0.1
PyJWT==2.9.0
pyproject_hooks==1.2.0
pywin32-ctypes==0.2.3
RapidFuzz==3.12.1
requests==2.32.3
requests-toolbelt==1.0.0
shellingham==1.5.4
sqlparse==0.5.3
tomlkit==0.13.2
trove-classifiers==2025.1.15.22
tzdata==2025.2
urllib3==2.3.0
userpath==1.9.2
virtualenv==20.29.1
